requirements-scheduler.txt
google-cloud-bigquery==3.2.0
google-cloud-storage==2.4.0
